// src/types.ts
export interface Notification {
  id: string;
  title: string;
  message: string;
  type: 'fee_request' | 'class_timing' | 'announcement' | 'general';
  targetType: 'individual' | 'class' | 'all';
  targetId?: string; // UID or Class Name
  senderId: string;
  senderName: string;
  createdAt: number;
  readBy: string[]; // Array of UIDs
}